requirements.txt
pandas
numpy
notebook
matplotlib
seaborn
scikit-learn
PyYAML
optuna
lightgbm
joblib
pytest